Finishing role admin

+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\admin\DashboardController;
 use App\Http\Controllers\admin\DataPemilihController;
 use App\Http\Controllers\admin\KandidatController;
+use App\Http\Controllers\admin\ManagementUser;
 use App\Http\Controllers\AuthentifikasiController;
 use App\Http\Middleware\SesiFalse;
 use Illuminate\Support\Facades\Route;
@@ -17,12 +18,19 @@ Route::middleware('role:admin')->group(function () {
         Route::get('dashboard', [DashboardController::class, 'index'])->name('dashboard.admin');
         Route::prefix('kandidat')->group(function () {
             Route::get('/', [KandidatController::class, 'show'])->name('admin.kandidat.store');
-            Route::post('/', [KandidatController::class, 'post']);
-            Route::put('{id}', [KandidatController::class, 'edit']);
-            Route::delete('{id}', [KandidatController::class, 'delete']);
+            Route::post('/', [KandidatController::class, 'post'])->name('admin.kandidat.post');
+            Route::put('{id}', [KandidatController::class, 'edit'])->name('admin.kandidat.edit');
+            Route::delete('{id}', [KandidatController::class, 'delete'])->name('admin.kandidat.delete');
         });
-        Route::prefix('data-pemilih')->group(function () {
-            Route::get('/', [DataPemilihController::class, 'show'])->name('admin.data-pemilih.store');
+        Route::get('data-pemilih', [DataPemilihController::class, 'show'])->name('admin.data-pemilih.store');
+        Route::prefix('user')->group(function () {
+            Route::get('/', [ManagementUser::class, 'show'])->name('admin.user.store');
+            Route::prefix('upload')->group(function () {
+                Route::post('/', [ManagementUser::class, 'post'])->name('admin.user.post');
+                Route::post('excel', [ManagementUser::class, 'postExcel'])->name('admin.user.excel');
+            });
+            Route::put('{id}', [ManagementUser::class, 'edit'])->name('admin.user.edit');
+            Route::delete('{id}', [ManagementUser::class, 'delete'])->name('admin.user.delete');
         });
     });
 });
